fixing small styles
