+++ src/pages/index.tsx
@@ -12,7 +12,7 @@ const Home: NextPage = () => {
       </Head>
 
       <div className="flex flex-col items-center justify-center h-screen space-y-6">
-        <Link href={"/pre_registration"}>Pre-Registration Page</Link>
+        <Link href={"/welcome"}>Pre-Registration Page</Link>
       </div>
     </>
   );
