+++ src/pages/welcome.tsx
@@ -1,17 +1,15 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Link from "next/link";
-import { useState } from "react";
-import NavBar from "../components/navigation";
+import NavBar from "../components/NavBar";
 import Background from "../components/background";
-
 import SocialButtons from "../components/SocialButtons";
 
 const PreRegistration: NextPage = () => {
   return (
     <>
       <Head>
-        <title>DH9 Pre-Registration Page</title>
+        <title>DH9 Welcome Page</title>
         <meta name="description" content="Generated by create-t3-app" />
         <link rel="icon" href="/favicon.ico" />
         <link
@@ -19,28 +17,27 @@ const PreRegistration: NextPage = () => {
           rel="stylesheet"
         />
       </Head>
-
-      <body className="font-main w-full min-h-screen overflow-x-hidden relative">
+      <div className="font-main w-full min-h-screen overflow-x-hidden relative">
         <NavBar />
         <Background />
         <main className="px-7 py-14 sm:px-14 md:pb-3 lg:pl-20 2xl:pt-20 md:absolute md:w-10/12 2xl:w-8/12">
-          <h1 className="text-2xl sm:text-3xl lg:text-5xl 2xl:text-6xl font-semibold text-white leading-tight">
+          <div className="text-2xl sm:text-3xl lg:text-5xl 2xl:text-6xl font-semibold text-white leading-tight">
             A weekend worth hacking,
             <br />@ DeltaHacks 9
-          </h1>
-          <p className="text-xl sm:text-2xl lg:text-3xl 2xl:text-4xl font-normal text-[#737373] pt-6 lg:pt-8 2xl:pt-10 lg:leading-tight">
+          </div>
+          <div className="text-xl sm:text-2xl lg:text-3xl 2xl:text-4xl font-normal text-[#737373] pt-6 lg:pt-8 2xl:pt-10 lg:leading-tight">
             At DeltaHacks, we believe change comes from dreaming big. Each year
             we enable over 800 students from across North America, working hard
             over 36 hours, to bring their big ideas to life. Unleash your
             creativity and make something great, we{"'"}ll handle the rest! Make
             big ideas a reality at DeltaHacks 9!
-          </p>
-          <p className="text-xl sm:text-2xl lg:text-3xl 2xl:text-4xl font-normal text-[#737373] pt-6 lg:pt-8 2xl:pt-10 lg:leading-tight">
-            If you have any questions, you can reach us at{" "}
+          </div>
+          <div className="text-xl sm:text-2xl lg:text-3xl 2xl:text-4xl font-normal text-[#737373] pt-6 lg:pt-8 2xl:pt-10 lg:leading-tight">
+            If you have any questions, you can <br />reach us at{" "}
             <a href="mailto: hello@deltahacks.com" className="text-sky-400">
               hello@deltahacks.com
             </a>
-          </p>
+          </div>
           <div className="flex flex-col md:flex-row gap-3 pt-6 lg:pt-8 2xl:pt-10">
             <Link href="#">
               <button className="btn btn-primary w-48 capitalize font-medium text-base">
@@ -54,10 +51,10 @@ const PreRegistration: NextPage = () => {
             </Link>
           </div>
         </main>
-        <footer className="sticky top-full grid place-items-end px-3 py-2 md:absolute md:top-0 md:bottom-0 md:right-5 md:w-2/12 2xl:w-4/12">
+        <footer className="sticky top-full grid place-items-end mb-3 px-2 py-2 md:absolute md:top-0 md:bottom-0 md:right-5 md:w-2/12 2xl:w-4/12">
           <SocialButtons />
         </footer>
-      </body>
+      </div>
     </>
   );
 };
